requirements.txt
instaloader>=4.9.5
python-dotenv>=1.0.0
tqdm>=4.64.0
selenium>=4.9.0
browser-cookie3>=0.18.0
biliup>=0.4.0
requests>=2.28.2
pytest>=7.3.1
pytest-mock>=3.10.0
